+++ plane/api/collections/__init__.py
@@ -1,0 +1,3 @@
+from .base import Collections
+
+__all__ = ["Collections"]
