+++ tests/unit/test_collections.py
@@ -28,31 +28,31 @@ class TestCollectionsAPI:
     """Test Collections CRUD."""
 
     def test_collection_crud(self, client: PlaneClient, workspace_slug: str) -> None:
-        created = client.collections.create_collection(
+        created = client.collections.create(
             workspace_slug, CreateCollection(name=_collection_name("SDK Collection"))
         )
         assert created is not None
         assert created.id is not None
 
         try:
-            collections = client.collections.list_collections(workspace_slug)
+            collections = client.collections.list(workspace_slug)
             assert isinstance(collections, list)
             assert any(c.id == created.id for c in collections)
 
-            retrieved = client.collections.retrieve_collection(workspace_slug, created.id)
+            retrieved = client.collections.retrieve(workspace_slug, created.id)
             assert isinstance(retrieved, Collection)
             assert retrieved.id == created.id
 
-            updated = client.collections.update_collection(
+            updated = client.collections.update(
                 workspace_slug,
                 created.id,
                 UpdateCollection(name=f"{created.name} (updated)", sort_order=25000),
             )
             assert updated.name == f"{created.name} (updated)"
         finally:
-            client.collections.delete_collection(workspace_slug, created.id, archive_pages=False)
+            client.collections.delete(workspace_slug, created.id, archive_pages=False)
 
-        collections_after = client.collections.list_collections(workspace_slug)
+        collections_after = client.collections.list(workspace_slug)
         assert not any(c.id == created.id for c in collections_after)
 
 
@@ -62,10 +62,10 @@ class TestCollectionPagesAPI:
     def test_add_list_search_move_and_remove_pages(
         self, client: PlaneClient, workspace_slug: str
     ) -> None:
-        source = client.collections.create_collection(
+        source = client.collections.create(
             workspace_slug, CreateCollection(name=_collection_name("Source Collection"))
         )
-        target = client.collections.create_collection(
+        target = client.collections.create(
             workspace_slug, CreateCollection(name=_collection_name("Target Collection"))
         )
         page = client.pages.create_workspace_page(
@@ -77,24 +77,24 @@ class TestCollectionPagesAPI:
         )
 
         try:
-            search_results = client.collections.search_addable_collection_pages(
+            search_results = client.collections.pages.search(
                 workspace_slug, source.id, search=page.name
             )
             assert any(r.id == page.id for r in search_results)
 
-            added = client.collections.add_collection_pages(
+            added = client.collections.pages.add(
                 workspace_slug, source.id, AddCollectionPages(page_ids=[page.id])
             )
             assert any(pc.page == page.id for pc in added)
 
-            listed = client.collections.list_collection_pages(workspace_slug, source.id)
+            listed = client.collections.pages.list(workspace_slug, source.id)
             assert isinstance(listed, PaginatedCollectionPageResponse)
             matching = [row for row in listed.results if row.page and row.page.get("id") == page.id]
             assert len(matching) == 1
             page_collection_id = matching[0].page_collection_id
             assert page_collection_id is not None
 
-            moved = client.collections.update_collection_page(
+            moved = client.collections.pages.update(
                 workspace_slug,
                 source.id,
                 page_collection_id,
@@ -102,8 +102,8 @@ class TestCollectionPagesAPI:
             )
             assert moved.collection == target.id
 
-            client.collections.remove_collection_page(workspace_slug, target.id, page_collection_id)
-            listed_after = client.collections.list_collection_pages(workspace_slug, target.id)
+            client.collections.pages.remove(workspace_slug, target.id, page_collection_id)
+            listed_after = client.collections.pages.list(workspace_slug, target.id)
             assert not any(
                 row.page and row.page.get("id") == page.id for row in listed_after.results
             )
@@ -114,9 +114,7 @@ class TestCollectionPagesAPI:
                 pass
             for collection in (source, target):
                 try:
-                    client.collections.delete_collection(
-                        workspace_slug, collection.id, archive_pages=False
-                    )
+                    client.collections.delete(workspace_slug, collection.id, archive_pages=False)
                 except Exception:
                     pass
 
@@ -124,7 +122,7 @@ class TestCollectionPagesAPI:
         self, client: PlaneClient, workspace_slug: str
     ) -> None:
         """Create page directly via CreatePage(collection_id=...), plus a sub-page."""
-        collection = client.collections.create_collection(
+        collection = client.collections.create(
             workspace_slug, CreateCollection(name=_collection_name("Parent Collection"))
         )
         parent_page = None
@@ -152,12 +150,12 @@ class TestCollectionPagesAPI:
 
             # The unfiltered listing returns only root-branch pages; sub-pages are
             # listed under their parent via the parent_id filter.
-            listed = client.collections.list_collection_pages(workspace_slug, collection.id)
+            listed = client.collections.pages.list(workspace_slug, collection.id)
             root_page_ids = {row.page.get("id") for row in listed.results if row.page}
             assert parent_page.id in root_page_ids
             assert child_page.id not in root_page_ids
 
-            children = client.collections.list_collection_pages(
+            children = client.collections.pages.list(
                 workspace_slug,
                 collection.id,
                 params=CollectionPageQueryParams(parent_id=parent_page.id),
@@ -172,9 +170,7 @@ class TestCollectionPagesAPI:
                     except Exception:
                         pass
             try:
-                client.collections.delete_collection(
-                    workspace_slug, collection.id, archive_pages=False
-                )
+                client.collections.delete(workspace_slug, collection.id, archive_pages=False)
             except Exception:
                 pass
 
@@ -189,18 +185,18 @@ class TestCollectionMembersAPI:
     """
 
     def test_collection_member_crud(self, client: PlaneClient, workspace_slug: str) -> None:
-        collection = client.collections.create_collection(
+        collection = client.collections.create(
             workspace_slug,
             CreateCollection(name=_collection_name("Private Collection"), access=1),
         )
         try:
-            members = client.collections.list_collection_members(workspace_slug, collection.id)
+            members = client.collections.members.list(workspace_slug, collection.id)
             assert isinstance(members, list)
             assert len(members) == 1
             owner_member = members[0]
             assert owner_member.access == 2  # EDIT, auto-assigned to the creator
 
-            updated_member = client.collections.update_collection_member(
+            updated_member = client.collections.members.update(
                 workspace_slug,
                 collection.id,
                 owner_member.id,
@@ -209,8 +205,6 @@ class TestCollectionMembersAPI:
             assert updated_member.access == 0
         finally:
             try:
-                client.collections.delete_collection(
-                    workspace_slug, collection.id, archive_pages=False
-                )
+                client.collections.delete(workspace_slug, collection.id, archive_pages=False)
             except Exception:
                 pass
